# docs/swagger.yaml
basePath: /
definitions:
...
  books.BookResponse:
    properties:
      author:
        type: string
      description:
        type: string
      id:
        type: integer
      isbn:
        type: string
      page_count:
        type: integer
      publication_year:
        type: integer
      title:
        type: string
    type: object
...
  inventory.AdjustSKUStockRequest:
    properties:
      change_by:
        type: integer
      comment:
        type: string
    type: object
  inventory.CreateSKURequest:
    properties:
      book_id:
        type: integer
      price_in_kopeks:
        minimum: 0
        type: integer
      stock_count:
        minimum: 0
        type: integer
      store_uuid:
        type: string
    required:
      - book_id
      - store_uuid
    type: object
  inventory.SKUResponse:
    properties:
      book_id:
        type: integer
      created_at:
        type: string
      price_in_kopeks:
        type: integer
      stock_count:
        type: integer
      store_id:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
    type: object
  inventory.SKUWithBookResponse:
    properties:
      book:
        $ref: '#/definitions/books.BookResponse'
      sku:
        $ref: '#/definitions/inventory.SKUResponse'
    type: object
  inventory.UpdateSKUPriceRequest:
    properties:
      new_price_in_kopeks:
        minimum: 0
        type: integer
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: { }
  description: Это REST API для сервиса сети книжных магазинов.
  title: Bookstores API
  version: "1.0"
paths:
...
  /skus:
    post:
      consumes:
        - application/json
      description: Создает новую товарную позицию (SKU), связывая книгу с магазином,
        ценой и остатком.
      parameters:
        - description: Данные для создания SKU
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/inventory.CreateSKURequest'
      produces:
        - application/json
      responses:
        "201":
          description: SKU успешно создан
          schema:
            $ref: '#/definitions/inventory.SKUResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Книга или магазин не найдены
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: SKU для этой книги в этом магазине уже существует
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Создать SKU
      tags:
        - skus
  /skus/{skuUUID}:
    get:
      description: Возвращает детальную информацию о SKU (включая данные о книге)
        по его UUID.
      parameters:
        - description: UUID товарной позиции (SKU)
          in: path
          name: skuUUID
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Информация о SKU и связанной книге
          schema:
            $ref: '#/definitions/inventory.SKUWithBookResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: SKU не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Получить SKU
      tags:
        - skus
  /skus/{skuUUID}/price:
    put:
      consumes:
        - application/json
      description: Устанавливает новую цену для существующей товарной позиции (SKU).
      parameters:
        - description: UUID товарной позиции (SKU)
          in: path
          name: skuUUID
          required: true
          type: string
        - description: Новая цена
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/inventory.UpdateSKUPriceRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Обновленный SKU
          schema:
            $ref: '#/definitions/inventory.SKUResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: SKU не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Обновить цену SKU
      tags:
        - skus
  /skus/{skuUUID}/stock-adjustments:
    post:
      consumes:
        - application/json
      description: Увеличивает или уменьшает количество товара на складе. Для уменьшения
        используйте отрицательное значение.
      parameters:
        - description: UUID товарной позиции (SKU)
          in: path
          name: skuUUID
          required: true
          type: string
        - description: Количество для изменения
          in: body
          name: input
          required: true
          schema:
            $ref: '#/definitions/inventory.AdjustSKUStockRequest'
      produces:
        - application/json
      responses:
        "200":
          description: Обновленный SKU
          schema:
            $ref: '#/definitions/inventory.SKUResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: SKU не найден
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Недостаточно товара для списания
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Скорректировать остатки
      tags:
        - skus
...
swagger: "2.0"